requirements.txt
praw
flask
psycopg2-binary